Add NumPy construction helpers (Wave 2)
Add the standard NumPy array-construction helpers, reusing the existing
runtime-dtype Array machinery (no new infra):

- arange(start, stop, step, dtype): the 2/3-arg form alongside the existing
  1-arg arange(n). Element count = max(0, ceil((stop-start)/step)); default
  dtype follows numpy (int64 if start/stop/step are all integers, else
  float64). Backed by a parallel affine ramp arange_range writing
  start + i*step on the NUMA workers.
- linspace(start, stop, num, endpoint, dtype): num evenly spaced points,
  float64 by default; the last point is forced exactly to stop when endpoint.
- eye(N, M, k, dtype) / identity(n, dtype): 2-D array with ones on the k-th
  diagonal; parallel for_loop over N*M cells.
- empty(shape, dtype): allocate without a value-fill (numpy.empty).
- zeros_like / ones_like / empty_like / full_like: pure-Python wrappers that
  reuse the existing constructors with a's shape and dtype.

Tests: tests/test_numpy_wave2.py (38 cases) checks values and dtype against a
numpy oracle for every helper. arange(n) now follows numpy's int64 default, so
a handful of existing tests that relied on the old float64 default request
dtype="float64" explicitly (the float arithmetic they exercise is unchanged).

check.sh green: ruff clean, 783 passed / 8 skipped, 100% coverage.

+++ tests/test_elementwise.py
@@ -34,7 +34,7 @@ def test_div_constant():
 
 
 def test_operator_matches_named_method():
-    a, b = hpx.arange(5000), hpx.full(5000, 2.0)
+    a, b = hpx.arange(5000, dtype="float64"), hpx.full(5000, 2.0)
     assert (a + b).sum() == a.add(b).sum()
     assert (a * b).sum() == a.mul(b).sum()
 
@@ -47,6 +47,6 @@ def test_empty():
 
 @pytest.mark.parametrize("op", ["add", "sub", "mul", "div"])
 def test_size_mismatch_raises(op):
-    a, b = hpx.arange(10), hpx.arange(11)
+    a, b = hpx.arange(10, dtype="float64"), hpx.arange(11, dtype="float64")
     with pytest.raises(ValueError):
         getattr(a, op)(b)
